pyproject.toml
[project]
name = "livekit-test2"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "weijay", email = "weijay0804@gmail.com"},
]
dependencies = ["livekit>=1.0.12", "livekit-api>=1.0.5", "faster-whisper>=1.2.0", "resampy>=0.4.3", "numpy>=2.2.6", "scipy>=1.16.1", "numba>=0.61.2", "fastapi>=0.116.1", "uvicorn>=0.35.0", "python-dotenv>=1.1.1"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
